a little rework in level_system. Add experience component, now level system component can be attached to weapons for level ups
