+++ Angless/Assets/Player/ExperienceSystemComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dba5d6e14bf89a24faeb55869213bc4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
