+++ Angless/Assets/Player/Player.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 483898756543579522}
   - component: {fileID: 483898756543579521}
   - component: {fileID: 483898756543579530}
+  - component: {fileID: 5062991553244883311}
   - component: {fileID: 483898756543579529}
   - component: {fileID: 483898756543579528}
   - component: {fileID: 483898756543579535}
@@ -131,10 +132,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f2b959c5e497ff245b3f7d20268f7168, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  requiredXPGrowthType: 0
-  requiredXP: 1
-  requiredXPGrowth: 1
-  currentXP: 0
   level: 1
   onLevelUp:
     m_PersistentCalls:
@@ -163,6 +160,37 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
+--- !u!114 &5062991553244883311
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 483898756543579527}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dba5d6e14bf89a24faeb55869213bc4e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  requiredXPGrowthType: 0
+  requiredXP: 1
+  requiredXPGrowth: 1
+  onXPReset:
+    m_PersistentCalls:
+      m_Calls:
+      - m_Target: {fileID: 483898756543579530}
+        m_TargetAssemblyTypeName: LevelSystemComponent, Assembly-CSharp
+        m_MethodName: LevelUp
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
+  currentXP: 0
 --- !u!114 &483898756543579529
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -192,6 +220,9 @@ MonoBehaviour:
   curHP: 0
   damageImmuneTimer: 0
   damageImmune: 0.5
+  OnDie:
+    m_PersistentCalls:
+      m_Calls: []
 --- !u!114 &483898756543579535
 MonoBehaviour:
   m_ObjectHideFlags: 0
